+++ ibm_botocore/__init__.py
@@ -26,7 +26,7 @@ import os
 import re
 import logging
 
-__version__ = '2.0.3.dev1'
+__version__ = '2.1.0'
 
 
 class NullHandler(logging.Handler):